requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
insightface==0.7.3
onnxruntime==1.16.3
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.4
gradio==4.16.0
